refactor(startup): move inkex and write-access guards into debug utils
